the landing page components are now able to be clicked on and go to the roadpages
