+++ src/app/(pages)/landing/CourseCard.js
@@ -3,6 +3,8 @@
 import React from 'react';
 import styles from './page.module.css';
 import ProgressBar from './progressBar';
+import { useRouter } from 'next/navigation'
+
 
 // course card component: individual course cards
 
@@ -14,10 +16,35 @@ function CourseCard({
     backgroundSize: 'cover',
     backgroundPosition: 'center',
   };
+  const router = useRouter()
+
+  const handleClick = () => {
+    console.log('Course card clicked!');
+    switch (name) {
+      case 'Plastic & Recycling':
+        router.push('/roadmap/plastic-and-recycling')
+        break;
+      case 'Eco-friendly Traveling':
+        router.push('/roadmap/eco-friendly-travel')
+        break;
+      case 'Ocean Pollution':
+        router.push('/roadmap/ocean-pollution')
+        break;
+      case 'Ecosystem Conservation':
+        router.push('/roadmap/conservation-and-restoration')
+        break;
+      case 'Climate Change':
+        router.push('/roadmap/climate-change')
+        break;
+      case 'Sustainability Lab':
+        router.push('/roadmap/sustainability-labs')
+        break;
+    }
+  };
 
   // code for cours name, duration, and progress bar
   return (
-    <div className={styles.courseCard} style={cardStyle}>
+    <div className={styles.courseCard} style={cardStyle} onClick={handleClick}>
       <div className={styles.courseName}>
         {name}
       </div>
