+++ src/app/(pages)/roadmap/plastic-and-recycling/page.jsx
@@ -1,5 +1,5 @@
 import Roadmap from "../roadmapfunction";
-import roadmapdata from "../roadmapdata"
+import roadmapdata from "../roadmap_data/roadmapdata"
 
 const key = 'plastic-and-recycling'; // for example
