+++ src/app/(pages)/roadmap/roadmap_data/roadmaps.js
@@ -1,0 +1,57 @@
+import course_1_step_1 from './course1/course_1_step_1.js'
+import course_1_step_2 from './course1/course_1_step_2.js'
+import course_1_step_3 from './course1/course_1_step_3.js'
+import course_1_step_4 from './course1/course_1_step_4.js'
+import course_1_step_5 from './course1/course_1_step_5.js'
+import course_1_step_6 from './course1/course_1_step_6.js'
+
+import course_2_step_1 from './course2/course_2_step_1.js'
+import course_2_step_2 from './course2/course_2_step_2.js'
+import course_2_step_3 from './course2/course_2_step_3.js'
+import course_2_step_4 from './course2/course_2_step_4.js'
+import course_2_step_5 from './course2/course_2_step_5.js'
+import course_2_step_6 from './course2/course_2_step_6.js'
+
+import course_3_step_1 from './course3/course_3_step_1.js'
+import course_3_step_2 from './course3/course_3_step_2.js'
+import course_3_step_3 from './course3/course_3_step_3.js'
+import course_3_step_4 from './course3/course_3_step_4.js'
+import course_3_step_5 from './course3/course_3_step_5.js'
+import course_3_step_6 from './course3/course_3_step_6.js'
+
+import course_4_step_1 from './course4/course_4_step_1.js'
+import course_4_step_2 from './course4/course_4_step_2.js'
+import course_4_step_3 from './course4/course_4_step_3.js'
+import course_4_step_4 from './course4/course_4_step_4.js'
+import course_4_step_5 from './course4/course_4_step_5.js'
+import course_4_step_6 from './course4/course_4_step_6.js'
+
+import course_5_step_1 from './course5/course_5_step_1.js'
+import course_5_step_2 from './course5/course_5_step_2.js'
+import course_5_step_3 from './course5/course_5_step_1.js'
+import course_5_step_4 from './course5/course_5_step_4.js'
+import course_5_step_5 from './course5/course_5_step_5.js'
+import course_5_step_6 from './course5/course_5_step_6.js'
+
+import course_6_step_1 from './course6/course_6_step_1.js'
+import course_6_step_2 from './course6/course_6_step_2.js'
+import course_6_step_3 from './course6/course_6_step_1.js'
+import course_6_step_4 from './course6/course_6_step_4.js'
+import course_6_step_5 from './course6/course_6_step_5.js'
+import course_6_step_6 from './course6/course_6_step_6.js'
+
+
+
+const coursesRoadmaps = {
+    course_1: [course_1_step_1, course_1_step_2, course_1_step_3, course_1_step_4, course_1_step_5, course_1_step_6],
+    course_2: [course_2_step_1, course_2_step_2, course_2_step_3, course_2_step_4, course_2_step_5, course_2_step_6],
+    course_3: [course_3_step_1, course_3_step_2, course_3_step_3, course_3_step_4, course_3_step_5, course_3_step_6],
+    course_4: [course_4_step_1, course_4_step_2, course_4_step_3, course_4_step_4, course_4_step_5, course_4_step_6],
+    course_5: [course_5_step_1, course_5_step_2, course_5_step_3, course_5_step_4, course_5_step_5, course_5_step_6],
+    course_6: [course_6_step_1, course_6_step_2, course_6_step_3, course_6_step_4, course_6_step_5, course_6_step_6]
+};
+
+
+
+
+export default coursesRoadmaps;
